Split GDB packet codec helpers
