+++ src/CMakeLists.txt
@@ -103,6 +103,7 @@ set(SRC_INTEGRATION
     ${APP_DIR}/integration/interpreter/execution.cpp
     ${APP_DIR}/integration/interpreter/breakpoints.cpp
     ${APP_DIR}/integration/gdb/gdbRemote.cpp
+    ${APP_DIR}/integration/gdb/gdbPacket.cpp
     ${APP_DIR}/integration/gdb/elfSymbols.cpp
     ${APP_DIR}/integration/keystone/assembler.cpp
 )
